requirements.txt
imbalanced-learn 
tensorflow
numpy
pandas
scikit-learn
optuna
pyyaml
python-dotenv
requests
beautifulsoup4
selenium
spektral
transformers
apache-beam[gcp]
google-cloud-bigquery
google-cloud-storage
tensorflow-probability
joblib
pytest
pytest-cov
mypy
pylint
flake8
black
isort
protobuf
tf_keras